+++ internal/provider/resource_coralogix_connector_test.go
@@ -16,6 +16,7 @@ package provider
 
 import (
 	"fmt"
+	"os"
 	"testing"
 
 	"github.com/google/uuid"
@@ -24,6 +25,9 @@ import (
 
 var (
 	connectorResourceName = "coralogix_connector.example"
+	msTeamsIntegrationId  = os.Getenv("MS_TEAMS_INTEGRATION_ID")
+	msTeamsTeamId         = os.Getenv("MS_TEAMS_TEAM_ID")
+	msTeamsChannelId      = os.Getenv("MS_TEAMS_CHANNEL_ID")
 )
 
 func TestAccCoralogixResourceGenericHttpsConnector(t *testing.T) {
@@ -205,6 +209,48 @@ func TestAccCoralogixResourcePagerdutyIncidentsConnector(t *testing.T) {
 					resource.TestCheckTypeSetElemNestedAttrs(connectorResourceName, "connector_config.fields.*", map[string]string{
 						"field_name": "service",
 						"value":      "PXXXXXX",
+					}),
+				),
+			},
+			{
+				ResourceName:      connectorResourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func TestAccCoralogixResourceMicrosoftTeamsConnector(t *testing.T) {
+	name := uuid.NewString()
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccRequiredEnvVarsPreCheck(t,
+				"MS_TEAMS_INTEGRATION_ID",
+				"MS_TEAMS_TEAM_ID",
+				"MS_TEAMS_CHANNEL_ID",
+			)
+		},
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccResourceCoralogixMicrosoftTeamsConnector(name),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(connectorResourceName, "id", name),
+					resource.TestCheckResourceAttr(connectorResourceName, "type", "microsoft_teams"),
+					resource.TestCheckResourceAttr(connectorResourceName, "name", name),
+					resource.TestCheckResourceAttr(connectorResourceName, "description", "test microsoft teams connector"),
+					resource.TestCheckTypeSetElemNestedAttrs(connectorResourceName, "connector_config.fields.*", map[string]string{
+						"field_name": "integrationId",
+						"value":      msTeamsIntegrationId,
+					}),
+					resource.TestCheckTypeSetElemNestedAttrs(connectorResourceName, "connector_config.fields.*", map[string]string{
+						"field_name": "teamId",
+						"value":      msTeamsTeamId,
+					}),
+					resource.TestCheckTypeSetElemNestedAttrs(connectorResourceName, "connector_config.fields.*", map[string]string{
+						"field_name": "channelId",
+						"value":      msTeamsChannelId,
 					}),
 				),
 			},
@@ -405,6 +451,31 @@ func testAccResourceCoralogixPagerdutyIncidentsConnector(name string) string {
      ]
    }
  }`, name, pagerDutyIntegrationId)
+}
+
+func testAccResourceCoralogixMicrosoftTeamsConnector(name string) string {
+	return fmt.Sprintf(`resource "coralogix_connector" "example" {
+   id               = "%[1]v"
+   type             = "microsoft_teams"
+   name             = "%[1]v"
+   description      = "test microsoft teams connector"
+   connector_config = {
+     fields = [
+       {
+         field_name = "integrationId"
+         value      = "%[2]v"
+       },
+       {
+         field_name = "teamId"
+         value      = "%[3]v"
+       },
+       {
+         field_name = "channelId"
+         value      = "%[4]v"
+       }
+     ]
+   }
+ }`, name, msTeamsIntegrationId, msTeamsTeamId, msTeamsChannelId)
 }
 
 func testAccResourceCoralogixEmailConnector(name string) string {